+++ intents/connectors/dialogflow_es/webhook.py
@@ -1,5 +1,5 @@
 """
-Here we manage webhook results
+Here we manage Dialogflow webhook results
 """
 import logging
 
@@ -7,7 +7,9 @@ from intents import FulfillmentResult, FulfillmentContext
 from intents.language_codes import ensure_language_code
 from intents.helpers.data_classes import to_dict
 from intents.connectors.dialogflow_es.names import event_name
+from intents.connectors.dialogflow_es.entities import MAPPINGS
 from intents.connectors.dialogflow_es import webhook_format as wf
+from intents.connectors.interface import serialize_intent_parameters
 
 logger = logging.getLogger(__name__)
 
@@ -17,7 +19,7 @@ def fulfillment_result_to_response(fulfillment_result: FulfillmentResult, contex
         followup_event = wf.WebhookResponseFollowupEvent(
             name=event_name(intent),
             languageCode=ensure_language_code(context.language).value,
-            parameters=intent.parameter_dict()
+            parameters=serialize_intent_parameters(intent, MAPPINGS)
         )
         return to_dict(wf.WebhookResponse(
             followupEventInput=followup_event