+++ intents/connectors/interface/__init__.py
@@ -18,6 +18,7 @@ specifically to:
 """
 from intents.connectors.interface.connector import Connector
 from intents.connectors.interface.entity_mappings import EntityMapping, \
-    StringEntityMapping, PatchedEntityMapping, ServiceEntityMappings, deserialize_intent_parameters
+    StringEntityMapping, PatchedEntityMapping, ServiceEntityMappings, \
+    deserialize_intent_parameters, serialize_intent_parameters
 from intents.connectors.interface.fulfillment import FulfillmentRequest, WebhookConfiguration
 from intents.connectors.interface.prediction import Prediction